Updated the XSD for the ORM configuration
Some properties were missing in the XSD and some others were available
only for the shortcut syntax and not for the expanded one, or the
opposite.
The ORM part of the XSD now uses groups to ensure that the shortcut and
expanded syntaxes are kept in sync (as done for the DBAL part already).