Allow tagging entity listeners
This allows tagging a service with an 'doctrine.orm.entity_listener'
tag. These tagged services will automatically be registered with the
entity listener resolver.

Note: The changes to the XML fixture cannot be parsed due to a bug in
the Symfony DependencyInjection component. A PR that fixes this issue
can be found at https://github.com/symfony/symfony/pull/9350. When that
PR has been merged, all tests in this commit pass.