Take a look if exists Namespace in the Classname
If we work with PostgreSQL Schema The Doctrine Generator may put Namespace in the Classname. The Namespace of the Entity represents the schema name for a table.

Right now, when we work with PostgreSQL Schema name 'access' and a table named 'user' and therefore therefore the Class name contain a string like 'Access\User' will be created a file named (on linux) 'Access\User.php'.