Enhance exception message
The error we want to improve occures if doctrine needs to establish
a connection to determine the platform version to register the types.
The idea is to provide some more information so the user can
understand what and why this is happening.

You can follow up on the discussions here:
https://github.com/doctrine/DoctrineBundle/issues/673
https://github.com/doctrine/DoctrineBundle/pull/677

Attention:
The composer doctrine/dbal dependency had to be updated to "^2.5.12".